Fix: issues with env file encryption leading to failed launch/depoloy

+++ cmd/launch.go
@@ -100,7 +100,11 @@ var launchCmd = &cobra.Command{
 		if utils.FileExists(fmt.Sprintf("./%s", envFileName)) {
 			hasEnvFile = true
 			pterm.Info.Printfln("Env file detected - Loading env vars from %s", envFileName)
-			utils.HandleEnvFile(envFileName, envVariables, &dockerEnvProperty, &envFileChecksum)
+			envHandleErr := utils.HandleEnvFile(envFileName, envVariables, &dockerEnvProperty, &envFileChecksum)
+			if envHandleErr != nil {
+				pterm.Error.Printfln("Something went wrong handling your env file: %s", envHandleErr)
+				os.Exit(1)
+			}
 			defer os.Remove("encrypted.env")
 		} else {
 			pterm.Info.Println("No env file detected - Skipping env parsing")
